+++ src/main/java/org/usf/jquery/web/RevisionIterator.java
@@ -53,23 +53,22 @@ public final class RevisionIterator implements Iterator<Entry<Integer, List<Year
 		throw new IllegalArgumentException("no revision");
 	}
 
-
 	static DBTable yearTable(String schema, String name, String tagname) {
 		return new DBTable(schema, name, tagname) {
 			@Override
-			public String sql(QueryParameterBuilder builder) {
-				return super.sql(builder) + "_" + currentRev.get().getKey();
+			public String sql(QueryParameterBuilder builder, String schema) {
+				return super.sql(builder, schema) + "_" + currentRev.get().getKey();
 			}
 		};
 	}
 
 	static DBColumn yearColumn() {
-		return constant(()-> currentRev.get().getKey().toString()); //get on build
+		return constant(()-> currentRev.get().getKey()); //get it on build
 	}
 	
 	static DBFilter monthFilter(DBColumn column) {
 		return b-> {
-			var values = currentRev.get().getValue();
+			var values = currentRev.get().getValue();  //get it on build
 			var filter = values.size() == 1 
 					? column.equal(values.get(0).getMonthValue())
 					: column.in(values.stream().map(YearMonth::getMonthValue).toArray(Integer[]::new));
